Fix incorrect page load on swipe pages.
Forcing the index of the swiper page interfers with the inner swiper state and causes unpredicted behavior. Now simply using the page state and passing that to the inner page for optimization logic to decide whether a re-render is necessary (we were using index which was not updated accurately, if I update index, it throws off the swiper component logic, so now I use page to determin whether a page needs to re-render).